requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
statsmodels>=0.12.0
matplotlib>=3.4.0
seaborn>=0.11.0
datasets>=2.0.0
together>=0.1.0
pyyaml>=6.0
backoff>=2.0.0
pytest>=6.2.0 